Highlighting nodes with linking and brushing.
Added a linking and brushing feature. When hovering over a node its size will be adjusted it, when the user no longer hover the node its size will return to its initial state. All nodes with the same character name in the different SVG:s are affected by this feature.